requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2016.9.26
cffi==1.15.1
charset-normalizer==2.0.12
clip @ git+https://github.com/openai/CLIP.git@3702849800aa56e2223035bccd1c6ef91c704ca8
cycler==0.11.0
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
decorator==5.1.1
defusedxml==0.7.1
einops==0.3.0
entrypoints==0.4
frozenlist==1.2.0
fsspec==2022.1.0
ftfy==6.0.3
future==0.18.2
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
Hydra==2.5
hydra-core==1.3.1
idna==3.4
idna-ssl==1.1.0
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipympl==0.8.8
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.2
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
kiwisolver==1.3.1
kornia==0.4.1
lpips==0.1.4
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
multidict==5.2.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
opencv-python==4.5.1.48
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-lightning==1.1.5
pytorch-ranger==0.1.1
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.24.2
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.12.1
test-tube==0.7.5
testpath==0.6.0
threadpoolctl==3.1.0
torch==1.10.0+cu111
torch-optimizer==0.3.0
torchaudio==0.10.0+rocm4.1
torchvision==0.11.0+cu111
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.6.1
yarl==1.7.2
zipp==3.6.0
